Add warp create command

+++ pcbridge-spigot/src/main/kotlin/com/projectcitybuild/features/warps/commands/warps/WarpDeleteCommand.kt
@@ -5,6 +5,7 @@ import com.projectcitybuild.support.messages.CommandHelpBuilder
 import com.projectcitybuild.support.spigot.BadCommandUsageException
 import com.projectcitybuild.support.spigot.CommandArgsParser
 import com.projectcitybuild.support.spigot.SpigotCommand
+import com.projectcitybuild.support.spigot.UnauthorizedCommandException
 import net.kyori.adventure.platform.bukkit.BukkitAudiences
 import net.kyori.adventure.text.Component
 import net.kyori.adventure.text.format.NamedTextColor
@@ -19,12 +20,15 @@ class WarpDeleteCommand(
     override val usage = CommandHelpBuilder() // TODO
 
     override suspend fun run(sender: CommandSender, args: Args) {
+        if (!sender.hasPermission("pcbridge.warp.manage")) {
+            throw UnauthorizedCommandException()
+        }
         warpRepository.delete(name = args.warpName)
 
-        val message = Component.text("${args.warpName} warp was deleted")
-            .color(NamedTextColor.GREEN)
-
-        audiences.sender(sender).sendMessage(message)
+        audiences.sender(sender).sendMessage(
+            Component.text("${args.warpName} warp deleted")
+                .color(NamedTextColor.GREEN)
+        )
     }
 
     data class Args(
